separated square generation from RunningGameState
